+++ src/index.ts
@@ -7,6 +7,7 @@ import IntroComponent from './components/intro/intro';
 import ReferenceComponent from './components/reference/reference';
 import AuthorsComponent from './components/authors/authors';
 import RulesComponent from './components/rules/rules';
+import SettingsComponent from './components/settings/settings';
 
 const router = new Router(document.querySelector('#root')!);
 function routePathsHandler() {
@@ -14,7 +15,7 @@ function routePathsHandler() {
     router.clearRoot();
 
     const diceBackgroundComponent =
-    new DiceBackgroundComponent(DiceTypes.BG_ALT);
+      new DiceBackgroundComponent(DiceTypes.BG_ALT);
     router.renderComponent(diceBackgroundComponent.render());
     diceBackgroundComponent.setup();
 
@@ -53,7 +54,7 @@ function routePathsHandler() {
     router.clearRoot();
 
     const diceBackgroundComponent =
-    new DiceBackgroundComponent(DiceTypes.BG_ALT);
+      new DiceBackgroundComponent(DiceTypes.BG_ALT);
     router.renderComponent(diceBackgroundComponent.render());
     diceBackgroundComponent.setup();
 
@@ -87,11 +88,11 @@ function routePathsHandler() {
     referenceComponentGame.setup();
     referenceComponentAuthors.setup();
     router.renderComponent(buttonWrapper);
-  } else if (router.checkPath('rules')){
+  } else if (router.checkPath('rules')) {
     router.clearRoot();
 
     const diceBackgroundComponent =
-    new DiceBackgroundComponent(DiceTypes.BG_ALT);
+      new DiceBackgroundComponent(DiceTypes.BG_ALT);
     router.renderComponent(diceBackgroundComponent.render());
     diceBackgroundComponent.setup();
 
@@ -129,7 +130,7 @@ function routePathsHandler() {
     router.clearRoot();
 
     const diceBackgroundComponent =
-    new DiceBackgroundComponent(DiceTypes.BG_ALT);
+      new DiceBackgroundComponent(DiceTypes.BG_ALT);
     router.renderComponent(diceBackgroundComponent.render());
     diceBackgroundComponent.setup();
 
@@ -163,6 +164,35 @@ function routePathsHandler() {
     referenceComponentGame.setup();
     referenceComponentAuthors.setup();
     router.renderComponent(buttonWrapper);
+
+  } else if (router.checkPath('game-settings')) {
+    router.clearRoot();
+
+    const diceBackgroundComponent =
+      new DiceBackgroundComponent(DiceTypes.BG_ALT);
+    router.renderComponent(diceBackgroundComponent.render());
+    diceBackgroundComponent.setup();
+
+    const diceBackgroundComponent2 = new DiceBackgroundComponent(DiceTypes.BG);
+    router.renderComponent(diceBackgroundComponent2.render());
+    diceBackgroundComponent2.setup();
+
+    const logoComponent = new LogoComponent();
+    router.renderComponent(logoComponent.render());
+    logoComponent.setup();
+
+    const settingsComponent = new SettingsComponent();
+    router.renderComponent(settingsComponent.render());
+    settingsComponent.setup();
+
+    const buttonWrapper = document.createElement('div');
+    buttonWrapper.setAttribute('class', 'menu-buttons');
+    buttonWrapper.appendChild(new ReferenceComponent('scores', 'scores').render());
+    buttonWrapper.appendChild(new ReferenceComponent('rules', 'rules').render());
+    buttonWrapper.appendChild(new ReferenceComponent('game-settings', 'game').render());
+    buttonWrapper.appendChild(new ReferenceComponent('authors', 'authors').render());
+    router.renderComponent(buttonWrapper);
+
   }
 }