+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added 
+### Added
 
 - Screens arrangement.
 - Game board Component.
 - Fabric.js.
+- Settings Component.
 - Modal Component.
 
 ### Changed
@@ -38,7 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Svg logo asset.
 - Types declaration for Typescript in `main.d.ts` file.
 - Github actions.
-
 
 ### Changed